refactor: modular architecture with thiserror and lib.rs
- Split domain.rs into domain/ module with focused submodules:
  - file_type.rs, file_entry.rs, decision.rs
  - app_state.rs, discovery.rs, decision_engine.rs
- Split tui.rs into tui/ module with submodules:
  - colors.rs, helpers.rs, input.rs
- Add error.rs with custom thiserror error types
- Add lib.rs exposing public library API
- Move PLAN.md to plans/ directory
- Update CLAUDE.md with new architecture documentation
