+++ src/tui/colors.rs
@@ -1,0 +1,10 @@
+use ratatui::style::Color;
+
+// Color scheme - Modern dark theme with vibrant accents
+pub const ACCENT_PRIMARY: Color = Color::Rgb(255, 107, 107); // Coral red for trash
+pub const ACCENT_SECONDARY: Color = Color::Rgb(107, 255, 158); // Mint green for keep
+pub const ACCENT_HIGHLIGHT: Color = Color::Rgb(255, 217, 102); // Golden yellow for highlights
+pub const TEXT_PRIMARY: Color = Color::Rgb(240, 240, 240); // Off-white
+pub const TEXT_SECONDARY: Color = Color::Rgb(160, 160, 170); // Muted gray
+pub const BG_DARK: Color = Color::Rgb(30, 30, 40); // Deep purple-black
+pub const BORDER_COLOR: Color = Color::Rgb(80, 80, 100); // Subtle purple-gray
