+++ src/tui/helpers.rs
@@ -1,0 +1,48 @@
+/// Calculates progress percentage
+pub fn calculate_progress(current: usize, total: usize) -> f64 {
+    if total == 0 {
+        0.0
+    } else {
+        (current as f64 / total as f64) * 100.0
+    }
+}
+
+/// Formats file size in human-readable format
+pub fn format_file_size(size: u64) -> String {
+    const KB: u64 = 1024;
+    const MB: u64 = KB * 1024;
+    const GB: u64 = MB * 1024;
+
+    if size >= GB {
+        format!("{:.1} GB", size as f64 / GB as f64)
+    } else if size >= MB {
+        format!("{:.1} MB", size as f64 / MB as f64)
+    } else if size >= KB {
+        format!("{:.1} KB", size as f64 / KB as f64)
+    } else {
+        format!("{} B", size)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_calculate_progress() {
+        assert_eq!(calculate_progress(0, 10), 0.0);
+        assert_eq!(calculate_progress(5, 10), 50.0);
+        assert_eq!(calculate_progress(10, 10), 100.0);
+        assert_eq!(calculate_progress(0, 0), 0.0);
+    }
+
+    #[test]
+    fn test_format_file_size() {
+        assert_eq!(format_file_size(0), "0 B");
+        assert_eq!(format_file_size(500), "500 B");
+        assert_eq!(format_file_size(1024), "1.0 KB");
+        assert_eq!(format_file_size(1536), "1.5 KB");
+        assert_eq!(format_file_size(1024 * 1024), "1.0 MB");
+        assert_eq!(format_file_size(1024 * 1024 * 1024), "1.0 GB");
+    }
+}
